examples/foc: move the motor controller code to separate files
